v2: disable spinners when not running in a tty
If oc-mirror is run behind a CI/CD pipeline or if output is redirected
to a file/pipe, the spinners' output is not shown and it looks like
oc-mirror is not making any progress. By using a regular log line
instead, we can still show some progress logs in those cases.